chore(*): Fix issue with FireFox not displaying vertial scrollbar. Remove hardcoded scrollbar heigth in hierarchical/tree grid.
